Check if returned exercise has changed

+++ src/app/pages/practitioner-exercise/practitioner-exercise.page.ts
@@ -34,7 +34,7 @@ export class PractitionerExercisePage {
             return;
         this.practitionerID = this.activatedRoute.snapshot.paramMap.get("id");
         let errorOccurred = false;
-        await this.practitionerService.GetPractitioner(this.practitionerID!).then(result => this.practitionerInfo = result).catch(() => errorOccurred = true);
+        await this.practitionerService.GetPractitioner(this.practitionerID!).then(result => this.practitionerInfo = result).catch(() => errorOccurred = true);//TODO from cache
         if (errorOccurred) {
             await this.alertService.ShowToast("Ocorreu um erro carregando as informações", undefined, "danger");
             return
@@ -103,15 +103,22 @@ export class PractitionerExercisePage {
                     rest: value.data.updatedWorkload!.rest,
                     load: value.data.updatedWorkload!.load
                 };
-                let errorOcurred = false;
-                await this.practitionerService.RemoveExercise(this.practitionerInfo!.exercisesID, oldExercise).catch(() => errorOcurred = true);
-                await this.practitionerService.AddExercise(this.practitionerInfo!.exercisesID, newExercise).catch(() => errorOcurred = true);
-                if (errorOcurred)
-                    await this.alertService.ShowToast("Não foi possível editar o exercício", undefined, "danger");
-                await this.populateExerciseList();
+                await this.updateEditedExercise(oldExercise, newExercise);
             }
         });
         await editExercisePopover.present();
+    }
+
+    async updateEditedExercise(oldExercise: Exercise, newExercise: Exercise) {
+        let errorOcurred = false;
+        if (oldExercise.series == newExercise.series && oldExercise.repetition == newExercise.repetition && oldExercise.rest == newExercise.rest && oldExercise.load == newExercise.load)
+            return;
+        await this.practitionerService.RemoveExercise(this.practitionerInfo!.exercisesID, oldExercise).catch(() => errorOcurred = true);
+        if (!errorOcurred)
+            await this.practitionerService.AddExercise(this.practitionerInfo!.exercisesID, newExercise).catch(() => errorOcurred = true);
+        if (errorOcurred)
+            await this.alertService.ShowToast("Não foi possível editar o exercício", undefined, "danger");
+        await this.populateExerciseList();
     }
 
     public async removeExerciseBtn(exercise: Exercise) {
@@ -125,7 +132,7 @@ export class PractitionerExercisePage {
         this.isLoading = false;
     }
 
-    private async populateExerciseList() {
+    private async populateExerciseList(fromCache = false) {//TODO from cache
         this.isLoading = true;
         this.practitionerExercises = await this.practitionerService.GetPractitionersExercises(this.practitionerInfo!.exercisesID);
         this.hasExercises = (typeof this.practitionerExercises != "undefined" && this.practitionerExercises.length > 0);
